requirements.txt
streamlit>=1.28.0
crewai>=0.30.0
crewai-tools>=0.2.0
python-dotenv>=1.0.0
groq>=0.4.0
